// src/interfaces/FormData.ts
export default interface IFormData {
  title: string;              
  description: string;        
  price: any;              
  state: string;              
  city: string;               
  isAvailable: boolean;       
  capacity: any;           
  bedrooms: number;          
  bathrooms: number;         
  photos: File[];             
}